backend/requirements.txt
# Backend Dependencies for ORBIT
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.1
httpx==0.26.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3

# AI / LLM
anthropic==0.18.0
tiktoken==0.5.2
langchain==0.1.4
langchain-anthropic==0.1.1

# Voice Processing
openai-whisper==20231117
soundfile==0.12.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.0
ruff==0.1.14
mypy==1.8.0
